strategy: split directional backtest lifecycle
