+++ src/strategy/directional_backtest.rs
@@ -13,22 +13,23 @@ use std::collections::HashMap;
 use std::fmt;
 
 use chrono::{DateTime, Utc};
-use rust_decimal::prelude::*;
 use rust_decimal::Decimal;
+use rust_decimal::prelude::*;
 use rust_decimal_macros::dec;
 use serde::{Deserialize, Serialize};
-use tracing::{debug, info, trace};
+use tracing::info;
 
 use crate::adapters::SpotPrice;
 use crate::domain::Side;
 use crate::strategy::backtest::BacktestResults;
 use crate::strategy::backtest_feed::{MarketFeed, UpdateType};
-use crate::strategy::backtest_recorder::{
-    BacktestRecorder, BacktestSignal, NullRecorder, PendingTrade, SignalType,
-};
+use crate::strategy::backtest_recorder::{BacktestRecorder, NullRecorder};
 use crate::strategy::execution_sim::ExecutionSimulator;
 use crate::strategy::fee_model::FeeModel;
 use crate::strategy::momentum::Direction;
+
+mod entry_lifecycle;
+mod position_lifecycle;
 
 // ─────────────────────────────────────────────────────────────
 // Config
@@ -397,592 +398,6 @@ impl DirectionalBacktestEngine {
 
         // Record equity curve
         self.record_equity(ts);
-    }
-
-    // ─── Entry logic (momentum + fair value + edge) ─────────
-
-    fn try_directional_entry(&mut self, symbol: &str, ts: DateTime<Utc>) {
-        // 1. Need: active events with S0, spot price history, PM asks
-        let windows: Vec<ActiveWindowInfo> = match self.active_events.get(symbol) {
-            Some(w) if !w.is_empty() => w.clone(),
-            _ => {
-                self.recorder.record_filtered(
-                    &BacktestSignal {
-                        signal_type: SignalType::Filtered,
-                        symbol: symbol.to_string(),
-                        direction: String::new(),
-                        timestamp: ts,
-                        p_hat: None,
-                        ev_net: None,
-                        sigma: None,
-                        market_price: None,
-                        spot_price: None,
-                        s0: None,
-                        time_remaining_secs: None,
-                        filter_reason: Some("no_active_event".to_string()),
-                        exit_reason: None,
-                        exit_price: None,
-                    },
-                    "no_active_event",
-                );
-                return;
-            }
-        };
-
-        // Shared preconditions: spot price with momentum, PM quotes
-        let (spot_price, momentum) = match self.spot_prices.get(symbol) {
-            Some(s) => {
-                // Use weighted momentum (10s/30s/60s) like live engine,
-                // fall back to 30s single-timeframe
-                let mom = s.weighted_momentum().or_else(|| s.momentum(30));
-                (s.price, mom)
-            }
-            None => {
-                self.recorder.record_filtered(
-                    &BacktestSignal {
-                        signal_type: SignalType::Filtered,
-                        symbol: symbol.to_string(),
-                        direction: String::new(),
-                        timestamp: ts,
-                        p_hat: None,
-                        ev_net: None,
-                        sigma: None,
-                        market_price: None,
-                        spot_price: None,
-                        s0: None,
-                        time_remaining_secs: None,
-                        filter_reason: Some("no_spot_data".to_string()),
-                        exit_reason: None,
-                        exit_price: None,
-                    },
-                    "no_spot_data",
-                );
-                return;
-            }
-        };
-
-        // Try entry on each active event window independently
-        for window in windows {
-            let (up_ask, down_ask) = self
-                .pm_asks_by_event
-                .get(&window.event_slug)
-                .copied()
-                .unwrap_or((None, None));
-            self.try_entry_for_window(symbol, ts, &window, spot_price, momentum, up_ask, down_ask);
-        }
-    }
-
-    /// Attempt entry on a specific event window using momentum-based fair value.
-    /// Mirrors the live MomentumDetector.check() → estimate_fair_value() → edge logic.
-    fn try_entry_for_window(
-        &mut self,
-        symbol: &str,
-        ts: DateTime<Utc>,
-        window: &ActiveWindowInfo,
-        st: Decimal,
-        momentum: Option<Decimal>,
-        up_ask: Option<Decimal>,
-        down_ask: Option<Decimal>,
-    ) {
-        // 2. Time remaining — must be within [min, max] window
-        let time_remaining = (window.end_time - ts).num_seconds() as f64;
-        if time_remaining <= 0.0 || time_remaining < self.config.min_time_remaining_secs as f64 {
-            return;
-        }
-        if time_remaining > self.config.max_time_remaining_secs as f64 {
-            return;
-        }
-
-        // 3. Momentum — need sufficient history
-        let momentum = match momentum {
-            Some(m) => m,
-            None => return, // insufficient price history
-        };
-
-        // 4. Minimum momentum threshold
-        if momentum.abs() < self.config.min_momentum {
-            self.recorder.record_filtered(
-                &BacktestSignal {
-                    signal_type: SignalType::Filtered,
-                    symbol: symbol.to_string(),
-                    direction: String::new(),
-                    timestamp: ts,
-                    p_hat: None,
-                    ev_net: None,
-                    sigma: None,
-                    market_price: None,
-                    spot_price: Some(st),
-                    s0: Some(window.s0),
-                    time_remaining_secs: Some(time_remaining),
-                    filter_reason: Some("momentum_below_threshold".to_string()),
-                    exit_reason: None,
-                    exit_price: None,
-                },
-                "momentum_below_threshold",
-            );
-            return;
-        }
-
-        // 5. Direction from momentum sign + select PM price
-        let (direction, market_ask) = if momentum > Decimal::ZERO {
-            match up_ask {
-                Some(ask) => (Direction::Up, ask),
-                None => return,
-            }
-        } else {
-            match down_ask {
-                Some(ask) => (Direction::Down, ask),
-                None => return,
-            }
-        };
-
-        // 6. Fair value estimation (sigmoid mapping, same as live engine)
-        let mut fair_value = Self::estimate_fair_value(momentum);
-
-        // 6b. Adjust for price_to_beat if enabled
-        if self.config.use_price_to_beat {
-            let time_remaining_secs = time_remaining as i64;
-            fair_value = Self::adjust_fair_value_for_price_to_beat(
-                fair_value,
-                momentum,
-                st,
-                window.s0,
-                time_remaining_secs,
-                window.end_time,
-            );
-        }
-
-        // 7. Price bounds check
-        if market_ask > self.config.max_entry_price || market_ask < self.config.min_entry_price {
-            self.recorder.record_filtered(
-                &BacktestSignal {
-                    signal_type: SignalType::Filtered,
-                    symbol: symbol.to_string(),
-                    direction: format!("{}", direction),
-                    timestamp: ts,
-                    p_hat: Some(fair_value.to_f64().unwrap_or(0.5)),
-                    ev_net: None,
-                    sigma: None,
-                    market_price: Some(market_ask),
-                    spot_price: Some(st),
-                    s0: Some(window.s0),
-                    time_remaining_secs: Some(time_remaining),
-                    filter_reason: Some("price_bounds".to_string()),
-                    exit_reason: None,
-                    exit_price: None,
-                },
-                "price_bounds",
-            );
-            return;
-        }
-
-        // 8. Edge = fair_value - pm_ask - fees
-        let best_bid = (market_ask - dec!(0.02)).max(dec!(0.01));
-        let depth_ratio = Decimal::from(self.config.shares_per_trade) / dec!(10000);
-        let cost = self
-            .fee_model
-            .all_in_cost(market_ask, best_bid, market_ask, depth_ratio);
-        let fee_per_share_usd = market_ask * cost.taker_fee;
-        let spread_plus_slip = cost.spread_cost + cost.depth_slippage;
-
-        let fair_value_f = fair_value.to_f64().unwrap_or(0.5);
-        let market_ask_f = market_ask.to_f64().unwrap_or(0.5);
-        let total_cost_f =
-            fee_per_share_usd.to_f64().unwrap_or(0.01) + spread_plus_slip.to_f64().unwrap_or(0.01);
-        let edge = fair_value_f - market_ask_f - total_cost_f;
-
-        if edge < self.config.entry_threshold {
-            self.recorder.record_filtered(
-                &BacktestSignal {
-                    signal_type: SignalType::Filtered,
-                    symbol: symbol.to_string(),
-                    direction: format!("{}", direction),
-                    timestamp: ts,
-                    p_hat: Some(fair_value_f),
-                    ev_net: Some(edge),
-                    sigma: None,
-                    market_price: Some(market_ask),
-                    spot_price: Some(st),
-                    s0: Some(window.s0),
-                    time_remaining_secs: Some(time_remaining),
-                    filter_reason: Some("edge_below_threshold".to_string()),
-                    exit_reason: None,
-                    exit_price: None,
-                },
-                "edge_below_threshold",
-            );
-            return;
-        }
-
-        // 9. Cooldown check
-        if let Some(last) = self.last_entry_time.get(symbol) {
-            let elapsed = (ts - *last).num_seconds();
-            if elapsed < self.config.cooldown_secs as i64 {
-                self.recorder.record_filtered(
-                    &BacktestSignal {
-                        signal_type: SignalType::Filtered,
-                        symbol: symbol.to_string(),
-                        direction: format!("{}", direction),
-                        timestamp: ts,
-                        p_hat: Some(fair_value_f),
-                        ev_net: Some(edge),
-                        sigma: None,
-                        market_price: Some(market_ask),
-                        spot_price: Some(st),
-                        s0: Some(window.s0),
-                        time_remaining_secs: Some(time_remaining),
-                        filter_reason: Some("cooldown".to_string()),
-                        exit_reason: None,
-                        exit_price: None,
-                    },
-                    "cooldown",
-                );
-                return;
-            }
-        }
-
-        // 10. Max positions check
-        if self.positions.len() >= self.config.max_concurrent_positions {
-            self.recorder.record_filtered(
-                &BacktestSignal {
-                    signal_type: SignalType::Filtered,
-                    symbol: symbol.to_string(),
-                    direction: format!("{}", direction),
-                    timestamp: ts,
-                    p_hat: Some(fair_value_f),
-                    ev_net: Some(edge),
-                    sigma: None,
-                    market_price: Some(market_ask),
-                    spot_price: Some(st),
-                    s0: Some(window.s0),
-                    time_remaining_secs: Some(time_remaining),
-                    filter_reason: Some("max_positions".to_string()),
-                    exit_reason: None,
-                    exit_price: None,
-                },
-                "max_positions",
-            );
-            return;
-        }
-
-        // 11. Don't enter if already holding same event+direction
-        let already_holding = self.positions.iter().any(|p| {
-            p.event_slug == window.event_slug
-                && std::mem::discriminant(&p.direction) == std::mem::discriminant(&direction)
-        });
-        if already_holding {
-            self.recorder.record_filtered(
-                &BacktestSignal {
-                    signal_type: SignalType::Filtered,
-                    symbol: symbol.to_string(),
-                    direction: format!("{}", direction),
-                    timestamp: ts,
-                    p_hat: Some(fair_value_f),
-                    ev_net: Some(edge),
-                    sigma: None,
-                    market_price: Some(market_ask),
-                    spot_price: Some(st),
-                    s0: Some(window.s0),
-                    time_remaining_secs: Some(time_remaining),
-                    filter_reason: Some("already_holding".to_string()),
-                    exit_reason: None,
-                    exit_price: None,
-                },
-                "already_holding",
-            );
-            return;
-        }
-
-        // 12. Execute entry via ExecutionSimulator
-        let sim_result =
-            self.execution_sim
-                .simulate_buy(market_ask, ts, self.config.shares_per_trade, 10_000);
-
-        let entry_cost = Decimal::from(sim_result.filled_shares) * sim_result.fill_price;
-        let entry_fee = self.fee_model.fee_shares(
-            Decimal::from(sim_result.filled_shares),
-            sim_result.fill_price,
-        ) * sim_result.fill_price;
-        let total_entry_cost = entry_cost + entry_fee;
-        if total_entry_cost > self.equity {
-            trace!(
-                "Skipping entry: insufficient equity ({} < {})",
-                self.equity,
-                total_entry_cost
-            );
-            return;
-        }
-
-        self.equity -= total_entry_cost;
-
-        self.positions.push(DirectionalPosition {
-            symbol: symbol.to_string(),
-            direction,
-            entry_price: sim_result.fill_price,
-            entry_time: ts,
-            shares: sim_result.filled_shares,
-            event_slug: window.event_slug.clone(),
-            s0: window.s0,
-            event_end_time: window.end_time,
-            entry_p_hat: fair_value_f,
-            entry_ev_net: edge,
-            entry_sigma: momentum.to_f64().unwrap_or(0.0), // store momentum in sigma field
-            latest_pm_price: market_ask,
-        });
-
-        self.last_entry_time.insert(symbol.to_string(), ts);
-
-        self.recorder.record_entry(&BacktestSignal {
-            signal_type: SignalType::Entry,
-            symbol: symbol.to_string(),
-            direction: format!("{}", direction),
-            timestamp: ts,
-            p_hat: Some(fair_value_f),
-            ev_net: Some(edge),
-            sigma: Some(momentum.to_f64().unwrap_or(0.0)),
-            market_price: Some(sim_result.fill_price),
-            spot_price: Some(st),
-            s0: Some(window.s0),
-            time_remaining_secs: Some(time_remaining),
-            filter_reason: None,
-            exit_reason: None,
-            exit_price: None,
-        });
-
-        debug!(
-            "ENTRY {} {} @ {:.4} | fv={:.3} edge={:.3} mom={:.4}%",
-            symbol,
-            direction,
-            sim_result.fill_price,
-            fair_value_f,
-            edge,
-            momentum.to_f64().unwrap_or(0.0) * 100.0
-        );
-    }
-
-    // ─── Fair value estimation (from live MomentumDetector) ───
-
-    /// Sigmoid-like mapping from momentum to fair value.
-    /// Mirrors `MomentumDetector::estimate_fair_value()` in momentum.rs.
-    fn estimate_fair_value(momentum: Decimal) -> Decimal {
-        let abs_momentum = momentum.abs();
-        let momentum_factor = if abs_momentum < dec!(0.001) {
-            // Very small moves: linear scaling (0.1% → 5%)
-            abs_momentum * dec!(50)
-        } else if abs_momentum < dec!(0.005) {
-            // Medium moves: moderate scaling (0.5% → ~21%)
-            dec!(0.05) + (abs_momentum - dec!(0.001)) * dec!(40)
-        } else {
-            // Large moves: diminishing returns (1% → ~36%)
-            dec!(0.21) + (abs_momentum - dec!(0.005)) * dec!(30)
-        };
-        // Cap at 90%
-        (dec!(0.50) + momentum_factor).min(dec!(0.90))
-    }
-
-    /// Adjust fair value based on distance to price_to_beat and time remaining.
-    /// Mirrors `MomentumDetector::estimate_fair_value_with_price_to_beat()`.
-    fn adjust_fair_value_for_price_to_beat(
-        base_fv: Decimal,
-        momentum: Decimal,
-        current_price: Decimal,
-        price_to_beat: Decimal,
-        time_remaining_secs: i64,
-        _end_time: DateTime<Utc>,
-    ) -> Decimal {
-        if price_to_beat <= Decimal::ZERO {
-            return base_fv;
-        }
-
-        let distance_pct = (current_price - price_to_beat) / price_to_beat;
-
-        // time_factor: fraction of time elapsed. Near expiry → time_factor → 1.0
-        let time_factor = (Decimal::ONE - Decimal::from(time_remaining_secs.max(0)) / dec!(900))
-            .max(Decimal::ZERO);
-
-        let direction_matches = (momentum > Decimal::ZERO && distance_pct > Decimal::ZERO)
-            || (momentum < Decimal::ZERO && distance_pct < Decimal::ZERO);
-
-        if direction_matches {
-            let boost = distance_pct.abs() * time_factor * dec!(0.5);
-            (base_fv + boost).min(dec!(0.95))
-        } else {
-            let reduction = distance_pct.abs() * dec!(0.3);
-            (base_fv - reduction).max(dec!(0.35))
-        }
-    }
-
-    // ─── Exit logic (directional, NOT arb) ───────────────────
-
-    fn check_exits(&mut self, ts: DateTime<Utc>) {
-        let mut to_close: Vec<(usize, Decimal, &str)> = Vec::new();
-
-        for (i, pos) in self.positions.iter().enumerate() {
-            let time_remaining = (pos.event_end_time - ts).num_seconds() as f64;
-
-            // A. Hold to settlement — no early exit at all. Binary options settle
-            // at $1.00 or $0.00, so unrealized PnL fluctuations are noise.
-            // The only meaningful exit is settlement itself.
-            // When enabled, skip ALL exit checks (time_stop, hard_stop) and wait
-            // for the EventState settlement event to close the position.
-            if self.config.hold_to_settlement {
-                continue;
-            }
-
-            // B. Time stop: <N secs remaining AND position is underwater.
-            //    Use unrealized PnL (market price vs entry price), NOT the probability model.
-            //    The prob model returns ~0.5 always, which would incorrectly exit winners
-            //    (pm_price > 0.5 → ev_now < 0 → false exit signal).
-            if time_remaining <= self.config.time_stop_secs as f64 && time_remaining > 0.0 {
-                let unrealized_per_share = pos.latest_pm_price - pos.entry_price;
-                if unrealized_per_share < Decimal::ZERO {
-                    to_close.push((i, pos.latest_pm_price, "time_stop"));
-                    continue;
-                }
-            }
-
-            // C. Hard stop: unrealized loss exceeds max
-            let unrealized = (pos.latest_pm_price - pos.entry_price) * Decimal::from(pos.shares);
-            if unrealized < Decimal::ZERO && unrealized.abs() > self.config.hard_stop_usd {
-                to_close.push((i, pos.latest_pm_price, "hard_stop"));
-                continue;
-            }
-        }
-
-        // Close in reverse order to preserve indices
-        to_close.sort_by(|a, b| b.0.cmp(&a.0));
-        for (idx, exit_price, reason) in to_close {
-            self.close_position(idx, exit_price, reason, ts);
-        }
-    }
-
-    // ─── Settlement ──────────────────────────────────────────
-
-    fn resolve_positions(
-        &mut self,
-        symbol: &str,
-        event_slug: &str,
-        up_won: bool,
-        ts: DateTime<Utc>,
-    ) {
-        let mut to_close = Vec::new();
-
-        for (i, pos) in self.positions.iter().enumerate() {
-            if pos.symbol == symbol && pos.event_slug == event_slug {
-                let exit_price = match (&pos.direction, up_won) {
-                    (Direction::Up, true) | (Direction::Down, false) => Decimal::ONE,
-                    _ => Decimal::ZERO,
-                };
-                to_close.push((i, exit_price));
-            }
-        }
-
-        to_close.sort_by(|a, b| b.0.cmp(&a.0));
-        for (idx, exit_price) in to_close {
-            self.close_position(idx, exit_price, "settlement", ts);
-        }
-    }
-
-    // ─── Close position ──────────────────────────────────────
-
-    fn close_position(&mut self, idx: usize, exit_price: Decimal, reason: &str, ts: DateTime<Utc>) {
-        let pos = self.positions.remove(idx);
-
-        // For settlement ($1 or $0), no need to simulate — it's binary payout.
-        // Fee at settlement ($1 or $0): p*(1-p) = 0, so settlement fee = $0.
-        // For early exits, simulate sell via ExecutionSimulator + exit fee.
-        let (final_price, proceeds, _exit_fee) = if reason == "settlement" {
-            let p = exit_price;
-            // At $1.00 or $0.00, the parabolic fee curve = 0 (p*(1-p) = 0)
-            (p, p * Decimal::from(pos.shares), Decimal::ZERO)
-        } else {
-            let sim_result = self
-                .execution_sim
-                .simulate_sell(exit_price, ts, pos.shares, 10_000);
-            let raw_proceeds = Decimal::from(sim_result.filled_shares) * sim_result.fill_price;
-            // Taker fee on sell: shares × price × feeRate × (p*(1-p))^exponent
-            let sell_fee = self.fee_model.fee_shares(
-                Decimal::from(sim_result.filled_shares),
-                sim_result.fill_price,
-            ) * sim_result.fill_price;
-            (sim_result.fill_price, raw_proceeds - sell_fee, sell_fee)
-        };
-
-        self.equity += proceeds;
-
-        // Entry fee was already deducted from equity at entry time,
-        // so PnL = proceeds - (shares × entry_price) already reflects the entry fee implicitly.
-        // But we also need to account for exit fee in the PnL.
-        let entry_fee = self
-            .fee_model
-            .fee_shares(Decimal::from(pos.shares), pos.entry_price)
-            * pos.entry_price;
-        let pnl = proceeds - Decimal::from(pos.shares) * pos.entry_price - entry_fee;
-        let holding_secs = (ts - pos.entry_time).num_seconds();
-
-        self.closed_trades.push(DirectionalClosedTrade {
-            symbol: pos.symbol.clone(),
-            direction: format!("{}", pos.direction),
-            entry_time: pos.entry_time,
-            exit_time: ts,
-            entry_price: pos.entry_price,
-            exit_price: final_price,
-            shares: pos.shares,
-            pnl,
-            won: pnl > Decimal::ZERO,
-            holding_secs,
-            exit_reason: reason.to_string(),
-            entry_p_hat: pos.entry_p_hat,
-            entry_ev_net: pos.entry_ev_net,
-            s0: pos.s0,
-            entry_sigma: pos.entry_sigma,
-        });
-
-        // Record exit signal and trade
-        self.recorder.record_exit(&BacktestSignal {
-            signal_type: SignalType::Exit,
-            symbol: pos.symbol.clone(),
-            direction: format!("{}", pos.direction),
-            timestamp: ts,
-            p_hat: Some(pos.entry_p_hat),
-            ev_net: Some(pos.entry_ev_net),
-            sigma: Some(pos.entry_sigma),
-            market_price: Some(final_price),
-            spot_price: None,
-            s0: Some(pos.s0),
-            time_remaining_secs: None,
-            filter_reason: None,
-            exit_reason: Some(reason.to_string()),
-            exit_price: Some(final_price),
-        });
-
-        self.recorder.record_trade(&PendingTrade {
-            symbol: pos.symbol,
-            direction: format!("{}", pos.direction),
-            entry_time: pos.entry_time,
-            exit_time: ts,
-            entry_price: pos.entry_price,
-            exit_price: final_price,
-            shares: pos.shares as i32,
-            pnl,
-            won: pnl > Decimal::ZERO,
-            holding_secs,
-            exit_reason: reason.to_string(),
-            entry_p_hat: Some(pos.entry_p_hat),
-            entry_ev_net: Some(pos.entry_ev_net),
-            entry_sigma: Some(pos.entry_sigma),
-            s0: Some(pos.s0),
-        });
-    }
-
-    /// Force-close remaining positions at latest PM price (data exhausted).
-    fn close_remaining_positions(&mut self) {
-        let ts = self.data_range_end.unwrap_or(Utc::now());
-        let indices: Vec<usize> = (0..self.positions.len()).rev().collect();
-        for idx in indices {
-            let price = self.positions[idx].latest_pm_price;
-            self.close_position(idx, price, "data_exhausted", ts);
-        }
     }
 
     // ─── Equity tracking ─────────────────────────────────────
